YCMBootstrap: Clean env variables before finding packages
If the environment variable CMAKE_PREFIX_PATH (or the variable PATH)
points to ${CMAKE_BINARY_DIR}/install subdirectory, the superbuild
will build the package the first time, but then, on the next run
find_package will find the package installed by the superbuild itself,
and therefore it will not try to build it.

This patch adds a step in the bootstrap script, in order to cleanup
the variables and avoid this issue.

Fixes: #21