ExternalProject: Fix SCM_DISCONNECTED patch
Always perform the skip-update target, forcing following steps to be
executed. This causes the configure, build, and install steps to be
executed and therefore changes to subprojects are now correctly
detected.

This also means that since the steps are now executed every time, build
takes more time and prints more unuseful output. The alternative is to
have a target to force the build when something changes, but it does not
seem to me a sufficiently generic solution.

Fixes #26