Properly handle absolute paths (hopefully this time for real)
HACK: configure_package_config_file uses a

  file(RELATIVE_PATH ... ... "${CMAKE_INSTALL_PREFIX}" )

to set the PACKAGE_PREFIX_DIR variable. This means that all the
variables in the build tree config file will use the absolute path, and
that the PACKAGE_PREFIX_DIR variable will contain the actual prefix
where the package is being installed.

This is probably not a problem, but in order to have only relative path
we temporarily save the CMAKE_INSTALL_PREFIX variable and replace it
with CMAKE_BINARY_DIR.