YCMEPHelper: Do not force unset CMake cache default arguments
When they are not set, this made impossible to set the values in a
subproject. For example if a subproject set BUILD_SHARED_LIBS as a
cached variable and in the main project this variable is not set, the
project was always built with BUILD_SHARED_LIBS="" (and therefore
disabled), and it was impossible to change this value.