YCMEPHelper: Replace <YCM_EP_CMAKE_ARGS> and do not configure the command strings
Replacing any variable can be dangerous and ExternalProject already
supports replacing <SOURCE_DIR> <BINARY_DIR> <INSTALL_DIR> and
<TMP_DIR>.

Instead of passing variables that are internal to YCMEPHelper, the user
should use  <SOURCE_DIR>, etc.

The <YCM_EP_CMAKE_ARGS> is an extra tag that is replaced with arguments
that YCMEPHelper would pass to the cmake configure command. Using this
it is possible to change the cmake command, but still pass the variables
required for a superbuild.

CC-Issue: #81