StandardFindModule: Add TARGET and REPLACE_TARGETS arguments
standard_find_module now handles the new `TARGET` and `REPLACE_TARGETS`
arguments, to create automatically the specified imported target and to
replace the known targets paths with the relative target name (in
pkg-config mode only).

If any of the `STANDARD_FIND_MODULE_USE_IMPORTED_TARGET` or
`STANDARD_FIND_MODULE_USE_IMPORTED_TARGET_<NAME>` are specified, and the
`TARGET` argument is given, the `<NAME>_LIBRARIES` variable content is
replaced with the corresponding imported target.